teste do servico anuncio

+++ Source/TA.Domain/Entity/IEntidade.cs
@@ -7,6 +7,6 @@ namespace TA.Domain.Entity
 {
     public interface IEntidade : IComparable<IEntidade>, IEquatable<IEntidade>
     {
-        int Id { get; set; }
+        ulong Id { get; set; }
     }
 }
